fix: clean up compiler warnings

+++ src/llama-model.cpp
@@ -621,7 +621,7 @@ struct ggml_backend_meta_split_state llama_meta_device_get_split_state(const str
                     return {{key_dim * (hparams.ssm_d_conv - 1), 2}, {value_dim * (hparams.ssm_d_conv - 1), 1}};
                 }
             } else {
-                const int64_t head_ratio = n_v_heads / n_k_heads;
+                const uint32_t head_ratio = static_cast<uint32_t>(n_v_heads / n_k_heads);
                 if (std::regex_match(tensor_name, pattern_qkv_weight) || std::regex_match(tensor_name, pattern_ssm_conv1d)) {
                     GGML_ASSERT(tensor->ne[axis] == 2*key_dim + value_dim);
                     return {{key_dim, 2 + head_ratio}};
